Add admin review moderation and decimal ratings to the Reviews API
Adds GET Reviews/All and DELETE Reviews/Items/{itemId}/Users/{userId},
both gated on Policies.RequiresElevation, so administrators can see
and remove any user's rating/comment, not just their own.

Also switches Rating from int to double across the DTOs/entry/manager
so ratings can carry fractional precision; UpsertReview truncates to
2 decimal places before persisting.

+++ MediaBrowser.Controller/Reviews/ReviewDto.cs
@@ -23,9 +23,10 @@ public class ReviewDto
     public string? UserName { get; set; }
 
     /// <summary>
-    /// Gets or sets the rating from 1-10, or null for a comment-only review.
+    /// Gets or sets the rating from 1-10, or null for a comment-only review. Precision is
+    /// truncated to 2 decimal places.
     /// </summary>
-    public int? Rating { get; set; }
+    public double? Rating { get; set; }
 
     /// <summary>
     /// Gets or sets the free-text comment, or null for a rating-only review.
