+++ MediaBrowser.Controller/Reviews/UpdateReviewRequest.cs
@@ -6,9 +6,10 @@ namespace MediaBrowser.Controller.Reviews;
 public class UpdateReviewRequest
 {
     /// <summary>
-    /// Gets or sets the rating from 1-10, or null to leave it unrated.
+    /// Gets or sets the rating from 1-10, or null to leave it unrated. Decimal values are
+    /// truncated to 2 decimal places.
     /// </summary>
-    public int? Rating { get; set; }
+    public double? Rating { get; set; }
 
     /// <summary>
     /// Gets or sets the comment text, or null/empty for no comment.
